ETL_pipeline/pipeline/etl_job/requirements.txt
pymongo
pandas
sqlalchemy
psycopg2-binary
vaderSentiment
python-dotenv